refactor: 사전 합의한대로 패키지 구조 변경 (#164)
