+++ src/main/java/com/sudo/railo/booking/application/ReservationService.java
@@ -21,10 +21,10 @@ import com.sudo.railo.booking.application.dto.response.SeatReservationDetail;
 import com.sudo.railo.booking.config.BookingConfig;
 import com.sudo.railo.booking.domain.PassengerSummary;
 import com.sudo.railo.booking.domain.Reservation;
-import com.sudo.railo.booking.domain.ReservationStatus;
+import com.sudo.railo.booking.domain.status.ReservationStatus;
 import com.sudo.railo.booking.exception.BookingError;
-import com.sudo.railo.booking.infra.ReservationRepository;
-import com.sudo.railo.booking.infra.ReservationRepositoryCustom;
+import com.sudo.railo.booking.infrastructure.reservation.ReservationRepository;
+import com.sudo.railo.booking.infrastructure.reservation.ReservationRepositoryCustom;
 import com.sudo.railo.global.exception.error.BusinessException;
 import com.sudo.railo.member.domain.Member;
 import com.sudo.railo.member.exception.MemberError;
