+++ src/main/java/com/sudo/railo/booking/domain/Ticket.java
@@ -4,6 +4,8 @@ import org.hibernate.annotations.Comment;
 import org.hibernate.annotations.OnDelete;
 import org.hibernate.annotations.OnDeleteAction;
 
+import com.sudo.railo.booking.domain.status.TicketStatus;
+import com.sudo.railo.booking.domain.type.PassengerType;
 import com.sudo.railo.global.domain.BaseEntity;
 import com.sudo.railo.train.domain.Seat;
