--- src/main/java/com/sudo/railo/booking/domain/TripType.java
@@ -1,5 +1,0 @@
-package com.sudo.railo.booking.domain;
-
-public enum TripType {
-	OW, RT
-}
