+++ src/main/java/com/sudo/railo/booking/domain/status/TicketStatus.java
@@ -1,4 +1,4 @@
-package com.sudo.railo.booking.domain;
+package com.sudo.railo.booking.domain.status;
 
 public enum TicketStatus {
 	ISSUED, USED, CANCELLED, REFUNDED
