+++ src/main/java/com/sudo/railo/booking/domain/type/PassengerType.java
@@ -1,4 +1,4 @@
-package com.sudo.railo.booking.domain;
+package com.sudo.railo.booking.domain.type;
 
 public enum PassengerType {
 	ADULT, CHILD, INFANT, SENIOR, DISABLED_HEAVY, DISABLED_LIGHT, VETERAN
