+++ src/main/java/com/sudo/railo/booking/domain/type/TripType.java
@@ -1,0 +1,5 @@
+package com.sudo.railo.booking.domain.type;
+
+public enum TripType {
+	OW, RT
+}
