+++ src/main/java/com/sudo/railo/booking/infrastructure/QrRepository.java
@@ -1,4 +1,4 @@
-package com.sudo.railo.booking.infra;
+package com.sudo.railo.booking.infrastructure;
 
 import org.springframework.data.jpa.repository.JpaRepository;
