+++ src/main/java/com/sudo/railo/booking/infrastructure/SeatReservationRepository.java
@@ -1,4 +1,4 @@
-package com.sudo.railo.booking.infra;
+package com.sudo.railo.booking.infrastructure;
 
 import java.util.Optional;
