+++ src/main/java/com/sudo/railo/booking/infrastructure/reservation/ReservationRepository.java
@@ -1,11 +1,11 @@
-package com.sudo.railo.booking.infra;
+package com.sudo.railo.booking.infrastructure.reservation;
 
 import java.time.LocalDateTime;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sudo.railo.booking.domain.Reservation;
-import com.sudo.railo.booking.domain.ReservationStatus;
+import com.sudo.railo.booking.domain.status.ReservationStatus;
 
 public interface ReservationRepository extends JpaRepository<Reservation, Long> {
 	void deleteAllByExpiresAtBeforeAndReservationStatusNot(LocalDateTime expiresAtBefore,
