+++ src/main/java/com/sudo/railo/booking/infrastructure/ticket/TicketRepository.java
@@ -1,4 +1,4 @@
-package com.sudo.railo.booking.infra;
+package com.sudo.railo.booking.infrastructure.ticket;
 
 import java.util.List;
